# docs/openapi.json
{
  "components": {
    "schemas": {
      "ApiError": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "additionalProperties": false,
        "properties": {
          "code": {
            "type": "string"
          },
          "hint": {
            "type": "string"
          },
          "issues": {
            "items": {
              "additionalProperties": false,
              "properties": {
                "message": {
                  "type": "string"
                },
                "path": {
                  "type": "string"
                }
              },
              "required": [
                "path",
                "message"
              ],
              "type": "object"
            },
            "type": "array"
          },
          "message": {
            "type": "string"
          },
          "request_id": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "type": "object"
      }
    },
    "securitySchemes": {
      "accessAssertion": {
        "description": "The signed Access JWT. Set by Cloudflare Access at the edge; the Worker verifies it itself rather than trusting that Access ran. Browsers may instead present the `CF_Authorization` cookie, which Cloudflare documents as not guaranteed to be passed in every context — the header is primary.",
        "in": "header",
        "name": "Cf-Access-Jwt-Assertion",
        "type": "apiKey"
      },
      "accessClientId": {
        "description": "An Access **service token** id. Presented with `CF-Access-Client-Secret`; Access exchanges the pair at the edge, so this header never reaches the Worker.",
        "in": "header",
        "name": "CF-Access-Client-Id",
        "type": "apiKey"
      },
      "accessClientSecret": {
        "description": "The matching Access service token secret. Required together with `CF-Access-Client-Id`.",
        "in": "header",
        "name": "CF-Access-Client-Secret",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "A self-hosted secrets manager on Cloudflare Workers and D1.\n\n**Slug alias routes.** Every environment-scoped route is served at two paths:\nthe canonical `/projects/{project}/environments/{env}/…` and the shorter\n`/p/{project}/e/{env}/…` for CLI ergonomics. Only the canonical form appears\nbelow; the alias is byte-identical in behaviour, matches **exactly** and never\nas a prefix.\n\n**There is no CORS.** This API emits no `Access-Control-Allow-Origin` and never\nwill. Omitting it is what stops another site reading a response in a victim's\nbrowser. The admin UI is same-origin and needs nothing.\n\n**`404` is used for absent and invisible alike.** A resource you cannot see is\nreported exactly as one that does not exist. The alternative leaks which project\nnames are in use to a caller with no grants at all.\n\n**Validation never echoes input.** A `422` body carries `issues` of\n`{path, message}`. The rejected value is dropped before the body is built, so a\nfailed secret write cannot put a plaintext value in a response, a log line or an\naudit row.",
    "title": "prick",
    "version": "0.0.0-dev"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/access/unknown-identities": {
      "get": {
        "description": "Read out of the audit log rather than from a table of pending requests, because there is nothing to pend: Access authenticates the caller at the edge and this Worker learns a subject exists at the moment it refuses it.\n\nSubjects that DO hold a live grant are excluded — they were denied something narrower than what they hold, which is ordinary authorization working rather than a missing introduction.\n\nThere is no id in these rows. Match `subject` against `GET /identities` to obtain the `identity_id` that `POST /grants` needs.",
        "operationId": "listUnknownIdentities",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "items": {
                    "additionalProperties": false,
                    "properties": {
                      "attempts": {
                        "maximum": 9007199254740991,
                        "minimum": 0,
                        "type": "integer"
                      },
                      "firstSeenAt": {
                        "maximum": 9007199254740991,
                        "minimum": 0,
                        "type": "integer"
                      },
                      "kind": {
                        "enum": [
                          "user",
                          "service"
                        ],
                        "type": "string"
                      },
                      "lastSeenAt": {
                        "maximum": 9007199254740991,
                        "minimum": 0,
                        "type": "integer"
                      },
                      "subject": {
                        "type": "string"
                      }
                    },
                    "required": [
                      "kind",
                      "subject",
                      "firstSeenAt",
                      "lastSeenAt",
                      "attempts"
                    ],
                    "type": "object"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Denied subjects, most recently seen first."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "`BAD_REQUEST` — a malformed `If-Match`, or a precondition this API cannot express."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "`UNAUTHENTICATED` — no valid Cloudflare Access assertion was presented."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "`FORBIDDEN` — authenticated, but no grant covers this scope. The denial is audited."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "`NOT_FOUND` — absent **or** invisible, deliberately indistinguishable. Returning 403 for one and 404 for the other would turn this API into an oracle for which slugs are in use."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "`VALIDATION_FAILED` — schema rejection, including an unknown field. `issues` carries `{path, message}` and never the submitted value."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "`INTERNAL`, `DECRYPT_FAILED`, `UNKNOWN_KID` or `SERVER_MISCONFIGURED`."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "`NO_ADMINS_CONFIGURED` or `IDENTITY_PROVIDER_UNAVAILABLE`."
          }
        },
        "summary": "Subjects that authenticated, were denied, and hold no grant (\"Seen but not granted\")",
        "tags": [
          "access"
        ]
      }
    },
...
    "/api/v1/grants/{id}": {
      "delete": {
        "description": "Refuses to remove the last global admin while `BOOTSTRAP_ADMINS` is empty. There is no recovery credential in this design — that is deliberate, and it means \"are you sure?\" is not a question a dialog can be trusted to ask. The only way back from an accidental lockout is editing a var and redeploying, so the operation is refused rather than confirmed.",
        "operationId": "revokeGrant",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Revoked."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "`BAD_REQUEST` — a malformed `If-Match`, or a precondition this API cannot express."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "`UNAUTHENTICATED` — no valid Cloudflare Access assertion was presented."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "`FORBIDDEN` — authenticated, but no grant covers this scope. The denial is audited."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "`NOT_FOUND` — absent **or** invisible, deliberately indistinguishable. Returning 403 for one and 404 for the other would turn this API into an oracle for which slugs are in use."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "`LAST_ADMIN` — this is the last usable global administrator and `BOOTSTRAP_ADMINS` is empty."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "`VALIDATION_FAILED` — schema rejection, including an unknown field. `issues` carries `{path, message}` and never the submitted value."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "`INTERNAL`, `DECRYPT_FAILED`, `UNKNOWN_KID` or `SERVER_MISCONFIGURED`."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "`NO_ADMINS_CONFIGURED` or `IDENTITY_PROVIDER_UNAVAILABLE`."
          }
        },
        "summary": "Revoke a grant",
        "tags": [
          "access"
        ]
      }
    },
...
    "/api/v1/identities": {
      "get": {
        "description": "Visible to **any** admin, at any scope — not global admin only. A project admin has to grant access within their project, and cannot do that without seeing who there is to grant it to; restricting this to global admins would make delegated administration decorative.\n\nAn identity row exists for every subject that has ever authenticated, including ones that were then denied: the row is written on the authenticated request itself, which is what makes a denied service token grantable at all.",
        "operationId": "listIdentities",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "items": {
                    "additionalProperties": false,
                    "properties": {
                      "disabled": {
                        "type": "boolean"
                      },
                      "displayName": {
                        "anyOf": [
                          {
                            "type": "string"
                          },
                          {
                            "type": "null"
                          }
                        ]
                      },
                      "id": {
                        "format": "uuid",
                        "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
                        "type": "string"
                      },
                      "kind": {
                        "enum": [
                          "user",
                          "service"
                        ],
                        "type": "string"
                      },
                      "lastSeenAt": {
                        "anyOf": [
                          {
                            "maximum": 9007199254740991,
                            "minimum": 0,
                            "type": "integer"
                          },
                          {
                            "type": "null"
                          }
                        ]
                      },
                      "subject": {
                        "type": "string"
                      }
                    },
                    "required": [
                      "id",
                      "kind",
                      "subject",
                      "displayName",
                      "disabled",
                      "lastSeenAt"
                    ],
                    "type": "object"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Identities, ordered by subject."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "`BAD_REQUEST` — a malformed `If-Match`, or a precondition this API cannot express."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "`UNAUTHENTICATED` — no valid Cloudflare Access assertion was presented."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "`FORBIDDEN` — authenticated, but no grant covers this scope. The denial is audited."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "`NOT_FOUND` — absent **or** invisible, deliberately indistinguishable. Returning 403 for one and 404 for the other would turn this API into an oracle for which slugs are in use."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "`VALIDATION_FAILED` — schema rejection, including an unknown field. `issues` carries `{path, message}` and never the submitted value."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "`INTERNAL`, `DECRYPT_FAILED`, `UNKNOWN_KID` or `SERVER_MISCONFIGURED`."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "`NO_ADMINS_CONFIGURED` or `IDENTITY_PROVIDER_UNAVAILABLE`."
          }
        },
        "summary": "List identities",
        "tags": [
          "access"
        ]
      }
    },
    "/api/v1/identities/{id}": {
      "patch": {
        "description": "Requires **global** admin, because `disabled` is a kill switch that outranks every grant at every scope — including `BOOTSTRAP_ADMINS`. A project admin flipping it would be revoking access to projects they have nothing to do with.\n\n`display_name` is not cosmetic: an access list of opaque `common_name` hex strings is unreadable, which is how a stale token survives three audits.",
        "operationId": "updateIdentity",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$schema": "https://json-schema.org/draft/2020-12/schema",
                "additionalProperties": false,
                "properties": {
                  "disabled": {
                    "type": "boolean"
                  },
                  "display_name": {
                    "anyOf": [
                      {
                        "maxLength": 128,
                        "minLength": 1,
                        "type": "string"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  }
                },
                "type": "object"
              }
            }
          },
          "description": "The fields to change.",
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$schema": "https://json-schema.org/draft/2020-12/schema",
                  "additionalProperties": false,
                  "properties": {
                    "disabled": {
                      "type": "boolean"
                    },
                    "displayName": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "id": {
                      "format": "uuid",
                      "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$",
                      "type": "string"
                    },
                    "kind": {
                      "enum": [
                        "user",
                        "service"
                      ],
                      "type": "string"
                    },
                    "lastSeenAt": {
                      "anyOf": [
                        {
                          "maximum": 9007199254740991,
                          "minimum": 0,
                          "type": "integer"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "subject": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "id",
                    "kind",
                    "subject",
                    "displayName",
                    "disabled",
                    "lastSeenAt"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "The updated identity."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "`BAD_REQUEST` — a malformed `If-Match`, or a precondition this API cannot express."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "`UNAUTHENTICATED` — no valid Cloudflare Access assertion was presented."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "`FORBIDDEN` — authenticated, but no grant covers this scope. The denial is audited."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "`NOT_FOUND` — absent **or** invisible, deliberately indistinguishable. Returning 403 for one and 404 for the other would turn this API into an oracle for which slugs are in use."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "`VALIDATION_FAILED` — schema rejection, including an unknown field. `issues` carries `{path, message}` and never the submitted value."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "`INTERNAL`, `DECRYPT_FAILED`, `UNKNOWN_KID` or `SERVER_MISCONFIGURED`."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "`NO_ADMINS_CONFIGURED` or `IDENTITY_PROVIDER_UNAVAILABLE`."
          }
        },
        "summary": "Rename or disable an identity",
        "tags": [
          "access"
        ]
      }
    },
...
  },
  "security": [
    {
      "accessClientId": [],
      "accessClientSecret": []
    },
    {
      "accessAssertion": []
    }
  ],
  "servers": [
    {
      "description": "This deployment.",
      "url": "/api/v1"
    }
  ],
...
}